Refactored MobileServiceRequestHelper into an abstract class and added functionality for calling APIs.
